Move NavigationServerManager out of server file
Moves NavigationServerManager out of server file.
