+++ modules/navigation_3d/3d/godot_navigation_server_3d.cpp
@@ -34,6 +34,7 @@
 
 #include "core/os/mutex.h"
 #include "scene/main/node.h"
+#include "scene/resources/3d/navigation_mesh_source_geometry_data_3d.h"
 
 using namespace NavigationDefaults3D;
