+++ modules/navigation_3d/register_types.cpp
@@ -46,6 +46,7 @@
 #include "core/object/callable_mp.h"
 #include "core/object/class_db.h"
 #include "servers/navigation_3d/navigation_server_3d.h"
+#include "servers/navigation_3d/navigation_server_3d_manager.h"
 
 #ifndef DISABLE_DEPRECATED
 NavigationMeshGenerator *_nav_mesh_generator = nullptr;
