+++ scene/3d/navigation/navigation_agent_3d.cpp
@@ -35,6 +35,7 @@
 #include "core/object/class_db.h"
 #include "scene/3d/navigation/navigation_link_3d.h"
 #include "scene/main/scene_tree.h"
+#include "scene/resources/mesh.h"
 #include "servers/navigation_3d/navigation_server_3d.h"
 #include "servers/rendering/rendering_server.h"
