+++ servers/navigation_2d/navigation_server_2d.h
@@ -30,13 +30,16 @@
 
 #pragma once
 
+#include "core/object/ref_counted.h"
+#include "core/os/rw_lock.h"
 #include "core/templates/rid_owner.h"
-#include "scene/resources/2d/navigation_mesh_source_geometry_data_2d.h"
-#include "scene/resources/2d/navigation_polygon.h"
-#include "servers/navigation_2d/navigation_path_query_parameters_2d.h"
-#include "servers/navigation_2d/navigation_path_query_result_2d.h"
+#include "core/variant/type_info.h"
 
 class Node;
+class NavigationPolygon;
+class NavigationMeshSourceGeometryData2D;
+class NavigationPathQueryParameters2D;
+class NavigationPathQueryResult2D;
 
 #ifdef CLIPPER2_ENABLED
 class NavMeshGenerator2D;
@@ -429,65 +432,6 @@ public:
 	void set_debug_navigation_avoidance_enable_obstacles_static(const bool p_value);
 	bool get_debug_navigation_avoidance_enable_obstacles_static() const;
 #endif // DEBUG_ENABLED
-};
-
-/// Manager used for the server singleton registration
-class NavigationServer2DManager : public Object {
-	GDCLASS(NavigationServer2DManager, Object);
-
-	static inline NavigationServer2DManager *singleton = nullptr;
-
-	struct ClassInfo {
-		String name;
-		Callable create_callback;
-
-		ClassInfo() {}
-
-		ClassInfo(const String &p_name, const Callable &p_create_callback) :
-				name(p_name),
-				create_callback(p_create_callback) {}
-
-		ClassInfo(const ClassInfo &p_ci) :
-				name(p_ci.name),
-				create_callback(p_ci.create_callback) {}
-
-		void operator=(const ClassInfo &p_ci) {
-			name = p_ci.name;
-			create_callback = p_ci.create_callback;
-		}
-	};
-
-	Vector<ClassInfo> navigation_servers;
-	int default_server_id = -1;
-	int default_server_priority = -1;
-
-	void on_servers_changed();
-
-protected:
-	static void _bind_methods();
-
-public:
-	static const String setting_property_name;
-
-	static NavigationServer2DManager *get_singleton();
-
-	void register_server(const String &p_name, const Callable &p_create_callback);
-	void set_default_server(const String &p_name, int p_priority = 0);
-	int find_server_id(const String &p_name);
-	int get_servers_count();
-	String get_server_name(int p_id);
-	NavigationServer2D *new_default_server();
-	NavigationServer2D *new_server(const String &p_name);
-
-	NavigationServer2DManager();
-	~NavigationServer2DManager();
-
-	static void initialize_server();
-	static void finalize_server();
-
-	static void initialize_server_manager();
-	static void finalize_server_manager();
-	static NavigationServer2D *create_dummy_server_callback();
 };
 
 VARIANT_ENUM_CAST(NavigationServer2D::ProcessInfo);
