+++ tests/servers/test_navigation_server_3d.cpp
@@ -40,7 +40,11 @@ TEST_FORCE_LINK(test_navigation_server_3d)
 #include "scene/3d/mesh_instance_3d.h"
 #include "scene/main/scene_tree.h"
 #include "scene/main/window.h"
+#include "scene/resources/3d/navigation_mesh_source_geometry_data_3d.h"
 #include "scene/resources/3d/primitive_meshes.h"
+#include "scene/resources/navigation_mesh.h"
+#include "servers/navigation_3d/navigation_path_query_parameters_3d.h"
+#include "servers/navigation_3d/navigation_path_query_result_3d.h"
 #include "servers/navigation_3d/navigation_server_3d.h"
 #include "tests/signal_watcher.h"
